group comparison file update
comments and descriptive added
