+++ PyEyeSim/_comparegroups.py
@@ -1,11 +1,21 @@
   
+# this file contains the EyeData methods for bewteen or within group comparisons.
+# not all functions can do both ---- yet  
+
 import numpy as np
 from numpy import matlib
 from scipy import stats,ndimage
 import pandas as pd
 import matplotlib.pyplot as plt
+
+
+# import  library helper functions. 
 from .statshelper import SaliencyMapFilt,SaccadesTrial,ScanpathL,StatEntropy
 from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim ,CheckCoor
+from .visualhelper import VisBinnedProg,PlotDurProg,JointBinnedPlot,MeanPlot,draw_ellipse,HistPlot
+
+
+
 
 def CompareGroupsFix(self,betwcond):
     '''
@@ -195,3 +205,35 @@ def FixDurProgGroups(self,withinColName,nfixmax=10):
         Err=stats.sem(np.nanmean(self.durprog[:,Idx],1),axis=0,nan_policy='omit')
         PlotDurProg(nfixmax,Y,Err,c)
     plt.legend()
+
+
+
+
+def BinnedDescriptivesGroups(self,withinColName):
+    ''' time-binned within trial descriptive progression, groups of stimuli'''
+    if hasattr(self,'binFixL')==False: 
+        print('run BinnedDescriptives first, than call this function for group wise visualization')
+    WhichC=self.GetCats(withinColName)
+    Colors=['navy','salmon','olive','orange','gray']
+    fig,ax=plt.subplots(nrows=3,ncols=1,figsize=(4,10))
+    for cc,c in enumerate(self.WithinConds):
+        Idx=np.nonzero(WhichC==c)[0]
+        axout=VisBinnedProg(self.tbins,np.nanmean(self.binFixL[:,Idx,:],1),'fixation duration (ms)',col=Colors[cc],label=c,axin=ax[0])
+        axout=VisBinnedProg(self.tbins,np.nanmean(self.saccadeAmp[:,Idx,:],1),'saccade ampl (pixel)',col=Colors[cc],label=c,axin=ax[1])
+        axout=VisBinnedProg(self.tbins,np.nanmean(self.totLscanpath[:,Idx,:],1),'scanpath length (pixel)',col=Colors[cc],label=c,axin=ax[2])
+        
+        ax1,ax2=JointBinnedPlot(self.tbins,np.nanmean(self.binFixL[:,Idx,:],1),np.nanmean(self.saccadeAmp[:,Idx,:],1),ylabel1='fixation duration (ms)',ylabel2='saccade ampl (pixel)')
+        ax1.set_title(c)
+        
+
+    ax[0].legend()
+    ax[1].legend()
+    ax[2].legend()
+    plt.tight_layout()      
+    
+    
+        
+    
+
+        
+
